WeatherInformation: current weather page

+++ WindowsPhone/WeatherInformation/WeatherInformation/Model/CurrentWeatherParser/Coord.cs
@@ -5,7 +5,7 @@ namespace WeatherInformation.Model.CurrentWeatherParser
     public class Coord
     {
         public double lon { get; set; }
-        public int lat { get; set; }
+        public double lat { get; set; }
     }
 }
