+++ WindowsPhone/WeatherInformation/WeatherInformation/Model/CurrentWeatherParser/Main.cs
@@ -5,7 +5,7 @@ namespace WeatherInformation.Model.CurrentWeatherParser
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
+        public double pressure { get; set; }
         public int humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
